fixed competence link, found bug in progress

+++ competence-gui/competence-webapp/src/uzuzjmd/competence/gui/client/tabs/ProgressTab.java
@@ -86,7 +86,7 @@ public class ProgressTab extends CompetenceTab {
 								.toMap(arg1);
 						for (String userName : userProgressMap.keySet()) {
 							progressPlaceHolder.add(new ProgressEntry(userName,
-									Double.valueOf(userProgressMap
+									Integer.valueOf(userProgressMap
 											.get(userName))));
 						}
 						alert = new Alert("Erfolgreich gefiltert",
